digital-metabolism-twin-public/requirements.txt
numpy==1.26.4
pandas==2.1.4
scikit-learn==1.3.2
xgboost==1.7.6
lightgbm==4.3.0
catboost==1.2.5
torch==2.1.2
onnxruntime==1.16.3
joblib==1.3.2
psutil==5.9.6
pyyaml==6.0.1
matplotlib==3.8.2
seaborn==0.13.1
scipy==1.11.4
shap==0.44.1
